+++ test/node/pipe.js
@@ -12,20 +12,39 @@ app.post('/', function(req, res){
 
 app.listen(3020);
 
-describe('request', function(){
-  describe('act as a writable stream', function(){
-    it('should format the url', function(done){
-      var req = request.post('http://localhost:3020')
-        , stream = fs.createReadStream('test/node/fixtures/user.json');
+describe('request pipe', function(){
+  afterEach(removeTmpfile);
 
-      req.type('json');
+  it('should act as a writable stream', function(done){
+    var req = request.post('http://localhost:3020');
+    var stream = fs.createReadStream('test/node/fixtures/user.json');
 
-      req.on('response', function(res){
-        res.body.should.eql({ name: 'tobi' });
-        done();
-      });
+    req.type('json');
 
-      stream.pipe(req);
-    })
+    req.on('response', function(res){
+      res.body.should.eql({ name: 'tobi' });
+      done();
+    });
+
+    stream.pipe(req);
   })
-})
+
+  it('should act as a readable stream', function(done){
+    var stream = fs.createWriteStream('test/node/fixtures/tmp.json');
+
+    var req = request.get('http://localhost:3025');
+    req.type('json');
+
+    req.on('end', function(){
+      JSON.parse(fs.readFileSync('test/node/fixtures/tmp.json', 'utf8')).should.eql({ name: 'tobi' });
+      done();
+    });
+    req.pipe(stream);
+  })
+});
+
+function removeTmpfile(done){
+  fs.unlink('test/node/fixtures/tmp.json', function(err){
+    done();
+  });
+}