+++ test/node/response-readable-stream.js
@@ -20,16 +20,18 @@ describe('response', function(){
         if (err) return done(err);
         var trackEndEvent = 0;
         var trackCloseEvent = 0;
+
         res.on('end',function(){
           trackEndEvent++;
+          trackEndEvent.should.equal(1);
+          trackCloseEvent.should.equal(0);  // close should not have been called
+          done();
         });
 
         res.on('close',function(){
           trackCloseEvent++;
-          trackEndEvent.should.equal(1);
-          trackCloseEvent.should.equal(1);
-          done();
         });
+
 
         (function(){ res.pause() }).should.not.throw();
         (function(){ res.resume() }).should.not.throw();