+++ test/server.js
@@ -97,10 +97,24 @@ app.get('/pets', function(req, res){
   res.send(['tobi', 'loki', 'jane']);
 });
 
+app.get('/text', function(req, res){
+  res.send("just some text");
+});
+
 app.get('/foo', function(req, res){
   res
     .header('Content-Type', 'application/x-www-form-urlencoded')
     .send('foo=bar');
+});
+
+app.post('/auth', function(req, res) {
+  var auth = req.headers.authorization,
+      parts = auth.split(' '),
+      credentials = new Buffer(parts[1], 'base64').toString().split(':'),
+      user = credentials[0],
+      pass = credentials[1];
+
+  res.send({ user : user, pass : pass });
 });
 
 app.use(express.static(__dirname + '/../'));